Missed typecast

+++ pgwatch2/metrics/wal_receiver/9.2/metric_standby.sql
@@ -1,4 +1,4 @@
 select
   (extract(epoch from now()) * 1e9)::int8 as epoch_ns,
-  pg_xlog_location_diff(pg_last_xlog_receive_location(), pg_last_xlog_replay_location()) as replay_lag_b,
+  pg_xlog_location_diff(pg_last_xlog_receive_location(), pg_last_xlog_replay_location())::int8 as replay_lag_b,
   extract(epoch from (now() - pg_last_xact_replay_timestamp()))::int8 as last_replay_s;
